add SetMediator method to Server interface
