+++ libcentrifugo/server/application.go
@@ -13,7 +13,6 @@ import (
 	"github.com/centrifugal/centrifugo/libcentrifugo/engine"
 	"github.com/centrifugal/centrifugo/libcentrifugo/metrics"
 	"github.com/centrifugal/centrifugo/libcentrifugo/proto"
-	"github.com/gorilla/securecookie"
 	"github.com/satori/go.uuid"
 )
 
@@ -60,8 +59,8 @@ type Application struct {
 	metrics *metrics.MetricsRegistry
 }
 
-// New returns new Application instance, the only required argument is
-// config, structure and engine must be set via corresponding methods.
+// New returns new server instance backed by Application, the only required
+// argument is config. Engine must be set via corresponding methods.
 func New(c *config.Config) (Server, error) {
 	app := &Application{
 		uid:        uuid.NewV4().String(),
@@ -830,57 +829,4 @@ func (app *Application) nClients() int {
 // nUniqueClients returns total amount of unique client connections to this node.
 func (app *Application) nUniqueClients() int {
 	return app.clients.nUniqueClients()
-}
-
-const (
-	// AuthTokenKey is a key for admin authorization token.
-	AuthTokenKey = "token"
-	// AuthTokenValue is a value for secure admin authorization token.
-	AuthTokenValue = "authorized"
-)
-
-func (app *Application) adminAuthToken() (string, error) {
-	app.RLock()
-	secret := app.config.AdminSecret
-	app.RUnlock()
-	if secret == "" {
-		logger.ERROR.Println("provide web_secret in configuration")
-		return "", ErrInternalServerError
-	}
-	s := securecookie.New([]byte(secret), nil)
-	return s.Encode(AuthTokenKey, AuthTokenValue)
-}
-
-// checkAdminAuthToken checks admin connection token which Centrifugo returns after admin login.
-func (app *Application) checkAdminAuthToken(token string) error {
-
-	app.RLock()
-	insecure := app.config.InsecureAdmin
-	secret := app.config.AdminSecret
-	app.RUnlock()
-
-	if insecure {
-		return nil
-	}
-
-	if secret == "" {
-		logger.ERROR.Println("provide admin_secret in configuration")
-		return ErrUnauthorized
-	}
-
-	if token == "" {
-		return ErrUnauthorized
-	}
-
-	s := securecookie.New([]byte(secret), nil)
-	var val string
-	err := s.Decode(AuthTokenKey, token, &val)
-	if err != nil {
-		return ErrUnauthorized
-	}
-
-	if val != AuthTokenValue {
-		return ErrUnauthorized
-	}
-	return nil
 }
